`MylistGroup.videos`
