+++ src/resolvers/MylistGroup/videos/Videos.graphql
@@ -1,0 +1,14 @@
+type MylistGroup {
+  videos(input: MylistGroupVideosInput!): [MylistGroupVideoAggregation!]!
+}
+
+input MylistGroupMylistsInput {
+  limit: Int! = 10
+  skip: Int! = 0
+  order: MylistGroupMylistsInputOrder! = { createdAt: DESC }
+}
+
+input MylistGroupMylistsInputOrder {
+  createdAt: SortOrder
+  updatedAt: SortOrder
+}
